requirements.txt
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.0
asttokens==2.2.0
attrs==22.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.4
bottle==0.12.23
Brlapi==0.8.3
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
cloudscraper==1.2.64
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
contourpy==1.0.6
cryptography==38.0.3
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
deap==1.3.3
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
dnspython==2.2.1
duplicity==0.8.21
ecdsa==0.18.0
evdev==1.4.0
executing==1.2.0
fastapi==0.92.0
fasteners==0.14.1
fastjsonschema==2.16.2
Flask==2.2.2
fonttools==4.38.0
fqdn==1.5.1
future==0.18.2
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
icmplib==3.0.3
idna==3.3
imageio==2.22.2
impacket==0.10.0
importlib-metadata==4.6.4
ipykernel==6.21.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.9.1
ldapdomaindump==0.9.3
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
maxminddb==2.2.0
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.2
mwparserfromhell==0.6.4
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.0.0
plotly==5.13.0
prometheus-client==0.16.0
prompt-toolkit==3.0.33
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.15.0
pycups==2.0.1
pydantic==1.10.5
pydot==1.4.2
pygame==2.1.2
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==2.4.7
pyRFC3339==1.1
PyRoxy @ git+https://github.com/MHProDev/PyRoxy.git@ea0f88dbc0573292ba5672124f69e4e7a31b544d
pyrsistent==0.19.3
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-jose==3.3.0
python-json-logger==2.0.4
python-magic==0.4.24
pytz==2022.1
pywikibot==7.7.3
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
reportlab==3.6.8
requests==2.27.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
seaborn==0.12.1
SecretStorage==3.3.1
Send2Trash==1.8.0
setproctitle==1.2.2
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.5.post1
sqlparse==0.4.2
stack-data==0.6.2
starlette==0.25.0
systemd-python==234
tenacity==8.2.1
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.6.0
typing_extensions==4.5.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.20.0
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wikipedia==1.4.0
xdg==5
xkit==0.0.0
yarl==1.8.1
zipp==1.0.0
